server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-telegram-bot==21.6
google-generativeai==0.8.3
groq==0.11.0
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
aiofiles==24.1.0
pillow==10.4.0
